refactored views files
