+++ src/frontend/src/views/NotFound.vue
@@ -1,0 +1,11 @@
+<script>
+export default {
+  name: "NotFound",
+  created() {
+    this.$router.push("/");
+  },
+  render() {
+    return "";
+  },
+};
+</script>
