+++ src/frontend/src/views/Orders.vue
@@ -29,9 +29,12 @@
 import { mapActions, mapState } from "vuex";
 import AppLayoutContent from "@/layouts/AppLayoutContent";
 import OrderCard from "@/modules/order/components/OrderCard";
+import { auth } from "@/middlewares";
 
 export default {
   name: "Orders",
+  layout: "AppLayoutMain",
+  middlewares: [auth],
   components: { AppLayoutContent, OrderCard },
   data() {
     return {
