--- src/frontend/src/views/ReadMe.md
@@ -1,1 +1,0 @@
-Данная директория содержит файлы представления (страницы)
