+++ src/frontend/src/views/profile/^create.vue
@@ -9,9 +9,12 @@
 <script>
 import { mapState, mapActions } from "vuex";
 import ProfileAddressForm from "@/modules/profile/components/ProfileAddressForm";
+import { auth } from "@/middlewares";
 
 export default {
   name: "ProfileCreate",
+  layout: "AppLayoutMain",
+  middlewares: [auth],
   components: {
     ProfileAddressForm,
   },
